Fix benchmark rates not using floats; add oreo and david's pc

+++ ereuse_devicehub/resources/event/schemas.py
@@ -299,7 +299,7 @@ class BenchmarkDataStorage(Benchmark):
 
 
 class BenchmarkWithRate(Benchmark):
-    rate = Integer(required=True)
+    rate = Float(required=True)
 
 
 class BenchmarkProcessor(BenchmarkWithRate):
